+++ lib/bcs.dart
@@ -1,5 +1,6 @@
-library bcs;
+library;
 
 export 'package:bcs_dart/index.dart';
+export 'package:bcs_dart/bcs_type.dart';
 export 'package:bcs_dart/utils.dart';
 export 'package:bcs_dart/legacy_bcs.dart';