+++ lib/hex.dart
@@ -1,9 +1,6 @@
-
 import 'dart:typed_data';
 
-
 class Hex {
-
   static String encode(List<int> bytes) {
     var buffer = StringBuffer();
     for (int byte in bytes) {
@@ -18,7 +15,11 @@ class Hex {
   static Uint8List decode(String hex) {
     String str = trimHex(hex);
     if (str.length % 2 != 0) {
-      str = "0" + str;
+      str = "0$str";
+    }
+
+    if (!RegExp(r'^[0-9a-fA-F]*$').hasMatch(str)) {
+      throw FormatException('Invalid hex string $hex');
     }
 
     var result = Uint8List(str.length ~/ 2);
@@ -37,7 +38,6 @@ class Hex {
 
   static String prefixHex(String hex) {
     if (hex.startsWith(RegExp('^0[xX]'))) return hex;
-    return "0x" + hex;
+    return "0x$hex";
   }
-
 }